solve some problem

+++ src/modules/order/order.validation.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 const orderValidationSchema = z.object({
     email: z.string().email({ message: "Invalid email address" }),
     productId: z.string(),
-    price: z.number().int().gt(0),
+    price: z.number().gt(0),
     quantity: z.number().int().gt(0),
 })
